+++ R/segregation.R
@@ -9,6 +9,7 @@
 #'
 #' @docType package
 #' @name segregation
+#' @keywords internal
 "_PACKAGE"
 
 #' @importFrom Rcpp sourceCpp
@@ -27,10 +28,10 @@ globalVariables(c(
     ".", "..base", "..fixed_margins", "..group", "..n_bootstrap", "..unit", "se", "stat",
     "M", "N_units", "i.freq1", "i.freq2", "iter", "ls1", "ls2", "p_unit_g_group1",
     "p_unit_g_group2", "pair", "pct_M",
-    "xmax", "xmin", "ymax", "ymin", "..cols", "p_overall",
+    "x", "xend", "y", "yend", "xmax", "xmin", "ymax", "ymin", "..cols", "p_overall",
     "freq_of", "freq_to",
     "cumul_prob_1", "cumul_prob_2", "group1", "group2", "pct_group_1",
-    ".data"
+    ".data", "new_unit", "old_unit"
 ))
 
 # log
@@ -179,6 +180,13 @@ prepare_data <- function(data, group, unit, weight, within = NULL) {
 
     if (!is.null(within)) {
         vars <- c(vars, within)
+    }
+
+    # drop unused factor levels - these can lead to problems downstream
+    for (var in vars) {
+        if (is.factor(data[[var]])) {
+            data[[var]] <- droplevels(data[[var]])
+        }
     }
 
     # collapse on vars, and select only positive weights