+++ inst/doc/plotting.R
@@ -13,16 +13,25 @@ knitr::opts_chunk$set(eval = BUILD_VIGNETTE)
 library("segregation")
 
 ## -----------------------------------------------------------------------------
-segcurve(subset(schools00, race %in% c("white", "black")),
+segcurve(subset(schools00, race %in% c("white", "asian")),
   "race", "school",
-  weight = "n"
+  weight = "n",
+  segment = "state" # leave this out to produce a single curve
 )
+
+## -----------------------------------------------------------------------------
+# converting to data.table makes this easier
+data.table::as.data.table(schools00)[
+  race %in% c("white", "asian"),
+  dissimilarity(.SD, "race", "school", weight = "n"),
+  by = .(state)
+]
 
 ## -----------------------------------------------------------------------------
 sch <- subset(schools00, state == "A")
 
 # basic segplot
-segplot(sch, "race", "school", weight = "n", axis_labels = "both")
+segplot(sch, "race", "school", weight = "n")
 
 # order by majority group (white in this case)
 segplot(sch, "race", "school", weight = "n", order = "majority")
@@ -39,6 +48,9 @@ segplot(sch, "race", "school",
   reference_distribution = ref
 )
 
+## -----------------------------------------------------------------------------
+segplot(sch, "race", "school", weight = "n", secondary_plot = "segregation")
+
 ## ---- results='hide'----------------------------------------------------------
 # compression based on window of 20 'neighboring' units
 # in terms of local segregation (alternatively, neighbors can be a data frame)
@@ -53,10 +65,14 @@ comp
 scree_plot(comp)
 
 ## -----------------------------------------------------------------------------
+dend <- as.dendrogram(comp)
+plot(dend, leaflab = "none")
+
+## -----------------------------------------------------------------------------
 sch_compressed <- merge_units(comp, n_units = 15)
 # or, for instance: merge_units(comp, percent = 0.80)
 head(sch_compressed)
 
 ## -----------------------------------------------------------------------------
-segplot(sch_compressed, "race", "school", weight = "n")
+segplot(sch_compressed, "race", "school", weight = "n", secondary_plot = "segregation")