+++ src/compression.cpp
@@ -14,7 +14,7 @@ struct CompressionResults
     std::vector<std::string> new_unit;
 };
 
-std::tuple<double, double> calculate_m(std::map<std::string, std::vector<double>> data)
+std::tuple<double, double> calculate_m(std::map<std::string, std::vector<double>> &data)
 {
     // create group sums
     int n_groups = data.begin()->second.size();
@@ -44,10 +44,51 @@ std::tuple<double, double> calculate_m(std::map<std::string, std::vector<double>
     return std::make_tuple(n_total, m_total);
 }
 
-std::tuple<double, double> calculate_twounit_m(std::vector<double> &unit1, std::vector<double> &unit2)
+std::vector<std::pair<std::string, double>> calculate_ls(std::map<std::string, std::vector<double>> &data)
 {
     // create group sums
-    int n_groups = unit1.size();
+    int n_groups = data.begin()->second.size();
+    std::vector<double> group_sums(n_groups, 0.0);
+    for (auto &[unit, counts] : data)
+    {
+        for (int i = 0; i < n_groups; i++)
+            group_sums[i] += counts[i];
+    }
+
+    // create group proportions
+    double n_total = std::accumulate(group_sums.begin(), group_sums.end(), 0);
+    std::vector<double> group_p(n_groups, 0.0);
+    for (int i = 0; i < n_groups; i++)
+    {
+        group_p[i] = group_sums[i] / n_total;
+    }
+
+    // create local segregation scores for each unit
+    std::vector<std::pair<std::string, double>> ls;
+    for (auto &[unit, counts] : data)
+    {
+        double n_unit = std::accumulate(counts.begin(), counts.end(), 0);
+        double ls_unit = 0.0;
+        for (int i = 0; i < n_groups; i++)
+        {
+            double p_group_given_unit = counts[i] / n_unit;
+            if (p_group_given_unit == 0)
+                continue;
+            ls_unit += p_group_given_unit * std::log(p_group_given_unit / group_p[i]);
+        }
+        ls.push_back({unit, ls_unit});
+    }
+
+    std::sort(ls.begin(), ls.end(), [](auto &left, auto &right)
+              { return left.second < right.second; });
+
+    return ls;
+}
+
+double calculate_reduction(double n, std::vector<double> &unit1, std::vector<double> &unit2)
+{
+    // create group sums
+    const int n_groups = unit1.size();
     double n_total = 0.0;
     std::vector<double> group_sums(n_groups, 0.0);
     for (int i = 0; i < n_groups; i++)
@@ -59,8 +100,8 @@ std::tuple<double, double> calculate_twounit_m(std::vector<double> &unit1, std::
     }
 
     // create unit sums
-    double n_unit1 = std::accumulate(unit1.begin(), unit1.end(), 0);
-    double n_unit2 = std::accumulate(unit2.begin(), unit2.end(), 0);
+    const double n_unit1 = std::accumulate(unit1.begin(), unit1.end(), 0);
+    const double n_unit2 = std::accumulate(unit2.begin(), unit2.end(), 0);
 
     // calculate M
     double m_total = 0.0;
@@ -82,45 +123,95 @@ std::tuple<double, double> calculate_twounit_m(std::vector<double> &unit1, std::
         }
     }
 
-    return std::make_tuple(n_total, m_total);
+    return n_total / n * m_total;
+}
+
+typedef std::pair<std::string, std::string> t_neighbor;
+
+t_neighbor neighbor_make_pair(std::string a, std::string b)
+{
+    if (a < b)
+        return t_neighbor(a, b);
+    else
+        return t_neighbor(b, a);
 }
 
 // [[Rcpp::export]]
 List compress_compute_cpp(
+    std::string neighbors_option,
     StringMatrix m_neighbors,
+    int n_neighbors,
     NumericMatrix m_data,
-    StringVector unit_names,
+    std::vector<std::string> unit_names,
     int max_iter)
 {
-    // prepare neighbors data structure (list of sets)
-    std::vector<std::set<std::string>> neighbors;
-    for (int i = 0; i < m_neighbors.nrow(); i++)
-    {
-        neighbors.push_back({});
-        for (int j = 0; j < m_neighbors.ncol(); j++)
-        {
-            neighbors[neighbors.size() - 1].insert(Rcpp::as<std::string>(m_neighbors(i, j)));
-        }
-    }
-
     // prepare main data structure: map, where the key is the unit name
     // and the values are the ordered group counts
     std::map<std::string, std::vector<double>> data;
     for (int i = 0; i < m_data.nrow(); i++)
     {
-        std::string unit = Rcpp::as<std::string>(unit_names[i]);
+        auto unit = unit_names[i];
         data[unit] = {};
         for (int j = 0; j < m_data.ncol(); j++)
-        {
             data[unit].push_back(m_data(i, j));
-        }
     }
 
     int n_groups = m_data.ncol();
-
     // compute total M index
     double n_total, m_total;
     std::tie(n_total, m_total) = calculate_m(data);
+
+    // prepare neighbors data structure (list of pairs, where order is unimportant)
+    std::map<t_neighbor, double> neighbors;
+    if (neighbors_option == "all")
+    {
+        for (int row = 0; row < unit_names.size(); row++)
+        {
+            for (int col = row + 1; col < unit_names.size(); col++)
+            {
+                neighbors[neighbor_make_pair(unit_names[row], unit_names[col])] = 0;
+            }
+        }
+    }
+    else if (neighbors_option == "local")
+    {
+
+        auto ls = calculate_ls(data);
+        for (int i = 0; i < ls.size(); i++)
+        {
+            for (int j = std::max(i - n_neighbors, 0);
+                 j < std::min(i + n_neighbors + 1, static_cast<int>(ls.size()) - 1);
+                 j++)
+            {
+                if (i != j)
+                    neighbors[neighbor_make_pair(ls[i].first, ls[j].first)] = 0;
+            }
+        }
+    }
+    else if (neighbors_option == "df")
+    {
+        for (int i = 0; i < m_neighbors.nrow(); i++)
+        {
+            std::string unit1 = Rcpp::as<std::string>(m_neighbors(i, 0));
+            std::string unit2 = Rcpp::as<std::string>(m_neighbors(i, 1));
+            if (unit1 != unit2)
+                neighbors[neighbor_make_pair(unit1, unit2)] = 0;
+        }
+    }
+
+    // calculate reduction for each neighbor pair
+    // (we don't do this in the previous step because otherwise we
+    // might do a lot of duplicate calculations)
+    for (const auto &[key, reduction] : neighbors)
+    {
+        neighbors[neighbor_make_pair(key.first, key.second)] = calculate_reduction(n_total, data[key.first], data[key.second]);
+    }
+
+    // determine maximum number of iterations
+    if (max_iter == -1)
+    {
+        max_iter = std::min(static_cast<int>(neighbors.size()), static_cast<int>(unit_names.size()) - 1);
+    }
 
     CompressionResults results;
     results.iter.reserve(max_iter);
@@ -138,59 +229,59 @@ List compress_compute_cpp(
     while (neighbors.size() > 0)
     {
         if (Progress::check_abort())
-            return List::create();
+            throw Rcpp::exception("user interruption");
 
-        // analyze reductions for all neighbors
+        // find smallest reduction
         double min_reduction = 10000;
-        int min_index;
-
-        for (int i = 0; i < neighbors.size(); i++)
+        t_neighbor min_key;
+        for (const auto &[key, reduction] : neighbors)
         {
-            auto iter = neighbors[i].begin();
-            double n_total_pair, m_total_pair;
-            std::tie(n_total_pair, m_total_pair) = calculate_twounit_m(
-                data[*iter], data[*next(iter)]);
-            // reduction = p_AB * M_AB
-            double reduction = n_total_pair / n_total * m_total_pair;
             if (reduction < min_reduction)
             {
                 min_reduction = reduction;
-                min_index = i;
+                min_key = key;
                 if (reduction == 0)
                     break;
             }
         }
 
-        auto iter = neighbors[min_index].begin();
-        const std::string unit_keep = *iter;
-        const std::string unit_delete = *next(iter);
-        // add counts of second to first, delete second
+        const std::string unit_keep = min_key.first;
+        const std::string unit_delete = min_key.second;
+        // add counts of 'delete' to 'keep', delete unit
         for (int i = 0; i < n_groups; i++)
-        {
             data[unit_keep][i] += data[unit_delete][i];
-        }
         data.erase(unit_delete);
-        // update all neighbor references
-        int mark_for_deletion;
-        for (int i = 0; i < neighbors.size(); i++)
+
+        // update neighbors
+        neighbors.erase(min_key);
+
+        std::vector<t_neighbor> delete_neighbors;
+        std::map<t_neighbor, double> new_neighbors;
+        for (const auto &[key, reduction] : neighbors)
         {
-            const bool is_in = neighbors[i].find(unit_delete) != neighbors[i].end();
-            if (is_in)
+            // update pairs if deleted unit is involved
+            if (key.first == unit_delete || key.second == unit_delete)
             {
-                neighbors[i].erase(unit_delete);
-                neighbors[i].insert(unit_keep);
+                // this is a pair some_unit - deleted_unit -- replace deleted_unit with unit_keep
+                delete_neighbors.push_back(key);
+                std::string some_unit = (key.first == unit_delete) ? key.second : key.first;
+                new_neighbors[neighbor_make_pair(unit_keep, some_unit)] =
+                    calculate_reduction(n_total, data[unit_keep], data[some_unit]);
             }
-            if (neighbors[i].size() == 1)
+            // recalculate reduction if kept unit is involved
+            else if (key.first == unit_keep || key.second == unit_keep)
             {
-                mark_for_deletion = i;
+                new_neighbors[key] = calculate_reduction(n_total, data[key.first], data[key.second]);
             }
         }
-        // clean up neighbors: erase equal elements
-        neighbors.erase(neighbors.begin() + mark_for_deletion);
 
-        // clean up neighbors: erase duplicates
-        std::sort(neighbors.begin(), neighbors.end());
-        neighbors.erase(std::unique(neighbors.begin(), neighbors.end()), neighbors.end());
+        // delete neighbors that involve the old unit
+        for (int i = 0; i < delete_neighbors.size(); i++)
+            neighbors.erase(delete_neighbors[i]);
+
+        // update and add new neighbors
+        for (const auto &[key, reduction] : new_neighbors)
+            neighbors[key] = reduction;
 
         // update results
         m_current -= min_reduction;
@@ -198,7 +289,15 @@ List compress_compute_cpp(
         counter += 1;
         results.iter.push_back(counter);
         results.M_wgt.push_back(min_reduction);
-        results.M.push_back(m_current);
+        if (n_units_current == 1)
+        {
+            // ensure that this is displayed as a true 0
+            results.M.push_back(0);
+        }
+        else
+        {
+            results.M.push_back(m_current);
+        }
         results.N_units.push_back(n_units_current);
         results.old_unit.push_back(unit_delete);
         results.new_unit.push_back(unit_keep);
@@ -216,4 +315,64 @@ List compress_compute_cpp(
         _["N_units"] = results.N_units,
         _["old_unit"] = results.old_unit,
         _["new_unit"] = results.new_unit);
+}
+
+int find_in_sets(std::string needle, std::vector<std::set<std::string>> haystack)
+{
+    for (int i = 0; i < haystack.size(); i++)
+    {
+        const bool is_in = haystack[i].find(needle) != haystack[i].end();
+        if (is_in == true)
+            return i;
+    }
+    return -1;
+}
+
+// [[Rcpp::export]]
+List get_crosswalk_cpp(std::vector<std::string> old_unit, std::vector<std::string> new_unit)
+{
+    std::vector<std::set<std::string>> bags;
+
+    for (int i = 0; i < old_unit.size(); i++)
+    {
+        int old_unit_bag = find_in_sets(old_unit[i], bags);
+        int new_unit_bag = find_in_sets(new_unit[i], bags);
+
+        if (old_unit_bag == -1 && new_unit_bag == -1)
+        {
+            // neither unit in bags - add new bag
+            bags.push_back({old_unit[i], new_unit[i]});
+        }
+        else if (old_unit_bag != -1 && new_unit_bag == -1)
+        {
+            // old_unit already exists - add new_unit to same bag
+            bags[old_unit_bag].insert(new_unit[i]);
+        }
+        else if (old_unit_bag == -1 && new_unit_bag != -1)
+        {
+            // new_unit already exists - add old_unit to same bag
+            bags[new_unit_bag].insert(old_unit[i]);
+        }
+        else if (old_unit_bag != -1 && new_unit_bag != -1)
+        {
+            // both units exist in different bags, merge the two
+            bags[old_unit_bag].insert(bags[new_unit_bag].begin(), bags[new_unit_bag].end());
+            bags.erase(bags.begin() + new_unit_bag);
+        }
+    }
+
+    // convert to List
+    List l(bags.size());
+    for (int i = 0; i < bags.size(); i++)
+    {
+        std::vector<std::string> bag(bags[i].size());
+        int index = 0;
+        for (auto el : bags[i])
+        {
+            bag[index] = el;
+            index++;
+        }
+        l[i] = bag;
+    }
+    return l;
 }