+++ tests/testthat/test_plots.R
@@ -6,8 +6,9 @@ library("segregation")
 context("plots")
 
 skip_if(!requireNamespace("ggplot2", quietly = TRUE))
+skip_if(!requireNamespace("patchwork", quietly = TRUE))
 
-plot_majority <- segplot(schools00, "race", "school", weight = "n", order = "majority")
+plot_majority <- segplot(schools00, "race", "school", weight = "n", order = "majority", hline = "white")
 plot_majority_fixed <- segplot(schools00, "race", "school", weight = "n", order = "majority_fixed")
 plot_seg <- segplot(schools00, "race", "school", weight = "n", order = "segregation")
 plot_entropy <- segplot(schools00, "race", "school",
@@ -50,16 +51,26 @@ test_that("reference", {
     expect_equal(sp$data[["unit"]], sp_ref$data[["unit"]])
 })
 
-test_that("axis_labels", {
-    left <- segplot(schools00, "race", "school", weight = "n", axis_labels = "left")
-    right <- segplot(schools00, "race", "school", weight = "n", axis_labels = "right")
-    both <- segplot(schools00, "race", "school", weight = "n", axis_labels = "both")
-    expect_equal(nrow(left$data), nrow(right$data))
-    expect_equal(nrow(both$data), nrow(right$data))
+test_that("secondary plot", {
+    plot_2a <- segplot(schools00, "race", "school",
+        weight = "n",
+        order = "segregation", secondary_plot = "segregation"
+    )
+    plot_2b <- segplot(schools00, "race", "school",
+        weight = "n",
+        order = "segregation", secondary_plot = "cumulative"
+    )
+
+    expect_equal(plot_seg$data, plot_2a[[1]]$data)
+    expect_equal(plot_seg$data, plot_2b[[1]]$data)
+    expect_true("patchwork" %in% class(plot_2a))
+    expect_true("patchwork" %in% class(plot_2b))
 })
 
 test_that("segcurve", {
     expect_error(segcurve(schools00, "race", "school", weight = "n"))
+    expect_error(segcurve(schools00, "race", "school", weight = "n", segment = c("a", "b")))
+    expect_error(segcurve(schools00, "race", "school", weight = "n", segment = c("state", "school")))
 
     p1 <- segcurve(subset(schools00, race %in% c("white", "black")),
         "race", "school",
@@ -69,4 +80,10 @@ test_that("segcurve", {
         "race", "school",
         weight = "n"
     )
+    p3 <- segcurve(subset(schools00, race %in% c("white", "asian")),
+        "race", "school",
+        weight = "n",
+        segment = "state"
+    )
+    expect_equal(p3$labels$colour, "state")
 })